Migrate theme setting from own localstorage to settingsStore
