+++ src/javascript/app/components/Layout/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import type { CSSPropertiesVars } from 'react';
 import { Outlet, Navigate, useMatches } from 'react-router';
 import Navigation from '../Navigation';
@@ -9,6 +9,8 @@ import PluginsContextProvider from '../../contexts/plugins/PluginsContextProvide
 import RemotePrinterContextProvider from '../../contexts/remotePrinter/RemotePrinterContextProvider';
 import { NavigationToolsProvider } from '../../contexts/navigationTools/NavigationToolsProvider';
 import { useScreenDimensions } from '../../../hooks/useScreenDimensions';
+import useSettingsStore from '../../stores/settingsStore';
+import { ThemeName } from '../../../consts/theme';
 
 import './index.scss';
 
@@ -19,6 +21,22 @@ export interface Handle {
 function Layout() {
   const matches = useMatches();
   const screenDimensions = useScreenDimensions();
+  const { themeName } = useSettingsStore();
+
+  useEffect(() => {
+    const classList = document.querySelector('html')?.classList;
+    if (!classList) {
+      return;
+    }
+
+    [ThemeName.BRIGHT, ThemeName.DARK].forEach((oldTheme) => {
+      if (oldTheme === themeName) {
+        classList.add(themeName);
+      } else {
+        classList.remove(oldTheme);
+      }
+    });
+  }, [themeName]);
 
   if (!matches[1]) {
     return <Navigate to="/gallery/page/1" replace />;
