+++ src/javascript/consts/theme.ts
@@ -1,0 +1,4 @@
+export enum ThemeName {
+  BRIGHT = 'theme-bright',
+  DARK = 'theme-dark',
+}
